pyproject.toml
[tool.poetry]
name = "tg-bot"
version = "0.1.0"
description = ""
authors = ["Vadim <bulayvadim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.8"
python-decouple = "^3.8"
freecurrencyapi = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
